Added package.json Gruntfile.js. Also updated README.md with instructions
Added some dependencies:
* Infusion
* Grunt and some plugins

From now on, we use the Grunt build system in order to set up everything.
In order to try out the extension, we build it, and the process is as it follows:

1. Concat and uglify the required dependencies (infusion and jQuery) and the extension/src/lib folder into a bundled .js file.
2. Copy all the required code and other files (manifest.json and extension/css folder) into the build folder.

With this, now we can try out the unpacked extension on Google Chrome.
Note that a build process is required after any update of the source code.