GPII-2466: Model being passed into UIO+ panel
Moved the primary model transformation up into chromeSettings to happen
when settings passed into the extension. In time, these transformations
may move to the solutions registry. Currently model values are being
passed into the context scripts and uio+ panel ( browser action popup ),
but are unidirectionaly. Need settings to be relayed back from the UIO+
panel to the applier and onto the context scripts and to chromeSettinsg
for enabling/disabling other extensions. Also need to update adjusters
and add new ones.