FLUID-6148: ToC working again.
There was an issue where if multiple content items were found. e.g. more 
than one article present on the page, that the levels subcomponent of 
the table of contents would get passed in multiple containers. Now we're 
taking into consideration that content might match on more than one 
element and only injecting the ToC container into one element. If 
multiple elements are found or none found, it renders into the container 
of the enactor. Which is likely the body.